rigth v 3.0

+++ Fillit/fillit.h
@@ -6,7 +6,7 @@
 /*   By: kroselin <kroselin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/28 14:19:21 by kroselin          #+#    #+#             */
-/*   Updated: 2019/11/20 13:17:38 by mdirect          ###   ########.fr       */
+/*   Updated: 2019/11/21 11:19:53 by mdirect          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@ int			correct_chars(char *buf);
 void		ft_print_bit(uint128_t n, int y);
 uint128_t	*resize_tetras(uint128_t *tetra, int y, int counter);
 uint128_t	count_lines(uint128_t *tetra);
-uint128_t	ft_move(uint128_t *tetra, int y);
+int			ft_move(uint128_t *tetra, int y);
 int 		func1(uint128_t map, uint128_t *tetra, int y);
 void		func2(uint128_t *tetra, int *y);
 void		ft_print_map(uint128_t *tetra, int y);
